worker/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
docker
radon
faiss-cpu
numpy
requests